fix linter errors and apply recommendations
