+++ src/index.js
@@ -1,18 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
-import { configureStore } from '@reduxjs/toolkit';
 import { Provider } from 'react-redux';
-import reducer from './redux/reducers';
 
 import './assets/stylesheets/reset.css';
 import './assets/stylesheets/index.css';
 import App from './reactjs/components/App';
+import store from './redux/configureStore';
 import reportWebVitals from './reportWebVitals';
-
-const store = configureStore({
-  reducer,
-});
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
