+++ src/redux/shared/actions.js
@@ -1,0 +1,2 @@
+// shared action
+export const RECEIVE_DATA = 'bookstore/shared/RECEIVE_DATA';
