+++ src/redux/shared/loading.js
@@ -1,6 +1,6 @@
-// loading.js
-import { RECEIVE_DATA } from '../actions/shared';
+import { RECEIVE_DATA } from './actions';
 
+// reducer
 export default function loading(state = true, action) {
   switch (action.type) {
     case RECEIVE_DATA:
